Update the modjk func module to instantiate the jk handle on each request
Before it was being instantiated once at module load time.  This is
bad because if for some reason modjkapi cannot contact the jkmanage
handler (for example during boot funcd starts before httpd so it
fails...)  then the module fails to deploy and a restart of funcd is
required to correct it.